Configuration install.sh

+++ install.sh
@@ -1,4 +1,22 @@
-#!/bin/bash
+#!/bin/zsh
+
+# Só para documentação de comandos
+# -d Verifica se uma pasta existe
+# -f Verifica se um arquivo existe
+# -z Verifica se a string está vazia
+# -s Verifica se existe o arquivo e se tem o tamanho > 0 bite
+#
+# $0 O nome do próprio script
+# $1, $2, $3... Mostra os argumentos passado em ordem
+# $# O número total de argumentos passado
+# $? Retorno do ultimo comando executado 0 = true; 1 = false
+#
+# zsh -x {SCRIPT} executa com debug
+# set -x ativa o debug dentro do script
+# set +x desativa o debug dentro do script
+#
+# 2>/dev/null redireciona o STDERR para o limbo
+
 
 # --- Configurações Iniciais ---
 DOTFILES_DIR="$HOME/dotfiles"
@@ -6,6 +24,8 @@ DOTFILES_DIR="$HOME/dotfiles"
 GREEN='\033[0;32m'
 BLUE='\033[0;34m'
 NC='\033[0m' # No Color
+ZSH_CUSTOM="${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"
+export NVM_DIR="$HOME/.nvm"
 
 echo -e "${BLUE}Iniciando a configuração do ambiente...${NC}"
 
@@ -73,87 +93,97 @@ link_dotfiles() {
 install_tools() {
     echo -e "${GREEN}Instalando ferramentas de desenvolvimento...${NC}"
 
-    # --- INSTALAÇÃO DO NEOVIM (Versão Github Release) ---
-    # Verifica se já está instalado ou se queremos forçar atualização
-    echo "Baixando Neovim (Latest Stable)..."
+    if ! command -v nvim >/dev/null; then
+      echo "${BLUE}Instalando Neovim...${NC}"
 
-    # 1. Baixa o tarball
-    curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
+      curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
 
-    # 2. Remove instalação anterior para evitar conflito
-    sudo rm -rf /opt/nvim-linux-x86_64
+      sudo rm -rf /opt/nvim-linux-x86_64
 
-    # 3. Extrai para /opt
-    sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
+      sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
 
-    # 4. Cria o link simbólico para o comando 'nvim' funcionar globalmente
-    # Se já existir um link antigo, remove primeiro
-    if [ -L /usr/local/bin/nvim ]; then
-        sudo rm /usr/local/bin/nvim
+      if [ -L /usr/local/bin/nvim ]; then
+          sudo rm /usr/local/bin/nvim
+      fi
+      sudo ln -s /opt/nvim-linux-x86_64/bin/nvim /usr/local/bin/nvim
+
+      rm nvim-linux-x86_64.tar.gz
+    else
+      echo "${BLUE}Neovim já está instalado.${NC}"
     fi
-    sudo ln -s /opt/nvim-linux-x86_64/bin/nvim /usr/local/bin/nvim
-
-    # 5. Limpa o arquivo baixado
-    rm nvim-linux-x86_64.tar.gz
-
-    echo "Neovim instalado com sucesso!"
 
     # --- INSTALAÇÃO DO SDKMAN (Java) ---
     if [ ! -d "$HOME/.sdkman" ]; then
-        echo "Instalando SDKMAN!..."
+        echo "${BLUE}Instalando SDKMAN!...${NC}"
         curl -s "https://get.sdkman.io" | bash
     else
-        echo "SDKMAN! já instalado."
+        echo "${BLUE}SDKMAN! já instalado."
     fi
 
     # --- INSTALAÇÃO DO OH MY ZSH ---
     if [ ! -d "$HOME/.oh-my-zsh" ]; then
-        echo "Instalando Oh My Zsh..."
+        echo "${BLUE}Instalando Oh My Zsh...${NC}"
 
         sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
     else
-        echo "Oh My Zsh já está instalado."
+        echo "${BLUE}Oh My Zsh já está instalado.${NC}"
     fi
 
     if [ ! -d "$HOME/.fzf" ]; then
-        echo "Instalando FZF..."
+        echo "${BLUE}Instalando FZF...${NC}"
 
         git clone --depth 1 https://github.com/junegunn/fzf.git "$HOME/.fzf"
         "$HOME/.fzf/install" --all
     else
-        echo "FZF já está instalado."
+        echo "${BLUE}FZF já está instalado.${NC}"
     fi
 
-    ZSH_CUSTOM="${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}"
-
     if [ ! -d "$ZSH_CUSTOM/themes/powerlevel10k" ]; then
-        echo "Baixando Powerlevel10k..."
+        echo "${BLUE}Baixando Powerlevel10k...${NC}"
 
         git clone --depth=1 https://github.com/romkatv/powerlevel10k.git "$ZSH_CUSTOM/themes/powerlevel10k"
     else
-        echo "Powerlevel10k já está instalado."
+        echo "${BLUE}Powerlevel10k já está instalado.${NC}"
     fi
 
     if [ ! -d "$ZSH_CUSTOM/plugins/zsh-autosuggestions" ]; then
-        echo "Baixando zsh-autosuggestions..."
+        echo "${BLUE}Baixando zsh-autosuggestions...${NC}"
 
         git clone https://github.com/zsh-users/zsh-autosuggestions "$ZSH_CUSTOM/plugins/zsh-autosuggestions"
     else
-        echo "zsh-autosuggestions já está instalado."
+        echo "${BLUE}zsh-autosuggestions já está instalado.${NC}"
     fi
 
     if [ ! -d "$ZSH_CUSTOM/plugins/zsh-syntax-highlighting" ]; then
-        echo "Baixando zsh-syntax-highlighting..."
+        echo "${BLUE}Baixando zsh-syntax-highlighting...${NC}"
 
         git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "$ZSH_CUSTOM/plugins/zsh-syntax-highlighting"
     else
-        echo "zsh-syntax-highlighting já está instalado."
+        echo "${BLUE}zsh-syntax-highlighting já está instalado.${NC}"
     fi
 
-    # TODO: Configurar a instalação do node
-    # curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.2/install.sh | bash
-    # \. "$HOME/.nvm/nvm.sh"
-    # nvm install 24
+    if [ ! -d "$NVM_DIR" ]; then
+        echo "${BLUE}Instalando o NVM...${NC}"
+
+        curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.2/install.sh | bash
+    else
+        echo "${BLUE}NVM já está instalado.${NC}"
+    fi
+
+
+    [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
+    if ! nvm ls 24 >/dev/null; then
+      echo "${BLUE}Instalando Node 24.${NC}"
+
+      nvm install 24
+      nvm use 24
+      nvm alias default 24
+    else
+      echo "${BLUE}Node 24 já está instalado.${NC}"
+
+      nvm use 24 >/dev/null
+      nvm alias default 24 >/dev/null
+    fi
 }
 
 # --- 4. Configuração Final ---
@@ -167,9 +197,9 @@ setup_shell() {
 
 # --- Execução ---
 ask_sudo
-install_packages
+# install_packages
 install_tools
-link_dotfiles
-setup_shell
+# link_dotfiles
+# setup_shell
 
 echo -e "${GREEN}Instalação concluída! Reinicie o terminal ou faça logout/login.${NC}"
